chore(ui), fix: add navigating to proper pages after form submittion, fix duplicate team name

+++ project/src/controllers/delete-project.controller.ts
@@ -1,4 +1,9 @@
-import { IRequest, ResponseCreator } from "@omniflow/common";
+import {
+    AnErrorOccurredError,
+    IRequest,
+    NotFoundError,
+    ResponseCreator,
+} from "@omniflow/common";
 import { IProjectRepository } from "../interfaces/repository.interface.js";
 
 export default function buildDeleteProjectController({
@@ -8,8 +13,18 @@ export default function buildDeleteProjectController({
 }) {
     return async (req: IRequest) => {
         const { currentProject } = req;
+        console.log({ currentProject });
 
-        console.log("delte project");
+        const toDelteProject = await projectRepository.get(currentProject._id);
+        console.log({ toDelteProject });
+
+        const projectDeleted = await projectRepository.delete(
+            currentProject._id
+        );
+
+        console.log({ projectDeleted });
+
+        if (!projectDeleted) throw new AnErrorOccurredError();
 
         const response = new ResponseCreator();
         await projectRepository.delete(currentProject._id);
